# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/me/audit/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Normal user can only view their own audit logs.\n普通用户只能查看自己的审计日志。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "List my audit logs / 查看我的审计日志",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page / 页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "page size / 每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start time (RFC3339 or YYYY-MM-DD) / 开始时间",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end time (RFC3339 or YYYY-MM-DD) / 结束时间",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-api_AuditLogPage"
                        }
                    }
                }
            }
        },
        "/api/v1/me/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "User changes their own password by providing old_password and new_password.\n用户提供旧密码与新密码，自助修改密码。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Change my password / 修改我的密码",
                "parameters": [
                    {
                        "description": "request / 请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
        "/api/v1/me/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List active web sessions (type=web, not revoked).\n列出有效 Web 会话（type=web 且未撤销）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "List my web sessions / 列出我的 Web 会话",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-array_api_TokenInfo"
                        }
                    }
                }
            }
        },
        "/api/v1/me/sessions/{jti}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Kick my session / 踢掉我的会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session jti / 会话标识",
                        "name": "jti",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
        "/api/v1/me/tokens": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List tokens (web sessions + api tokens) of current user.\n列出当前用户的 Token（Web 会话 + API Token）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "List my tokens / 列出我的 Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-array_api_TokenInfo"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a permanent API token for the current user.\n为当前用户创建永久 API Token。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Create my API token / 创建我的 API Token",
                "parameters": [
                    {
                        "description": "request / 请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateMyTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-api_CreateAPITokenResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Self-service revoke all tokens without admin.\n普通用户自助撤销自己所有 Token（不走管理员）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Revoke all my tokens / 撤销我所有 Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
        "/api/v1/me/tokens/{jti}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Revoke my token / 撤销我的指定 Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token jti / token 标识",
                        "name": "jti",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.AuditLogItem": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.AuditLogPage": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.AuditLogItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.ChangePasswordRequest": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "api.CreateAPITokenResponse": {
            "type": "object",
            "properties": {
                "jti": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.CreateMyTokenRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "api.TokenInfo": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "idle_timeout_seconds": {
                    "type": "integer"
                },
                "issued_at": {
                    "type": "string"
                },
                "jti": {
                    "type": "string"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.TokenType"
                }
            }
        },
...
        "models.TokenType": {
            "type": "string",
            "enum": [
                "web",
                "api"
            ],
            "x-enum-varnames": [
                "TokenTypeWeb",
                "TokenTypeAPI"
            ]
        },
        "response.Envelope-any": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {},
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
...
        "response.Envelope-api_AuditLogPage": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "$ref": "#/definitions/api.AuditLogPage"
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
        "response.Envelope-api_CreateAPITokenResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "$ref": "#/definitions/api.CreateAPITokenResponse"
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
...
        "response.Envelope-array_api_TokenInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TokenInfo"
                    }
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
        "response.ErrorCode": {
            "type": "integer",
            "enum": [
                0,
                10000,
                10001,
                10002,
                10003,
                10004,
                10005
            ],
            "x-enum-varnames": [
                "CodeOK",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeConflict",
                "CodeInternal"
            ]
        }
    }
}